entered search icon and main search functionality added
