+++ src/styles/css/styles.css
@@ -650,7 +650,7 @@
   width: 100%;
 }
 .ant-input-prefix {
-  margin-top: 1rem;
+  margin-top: 0.5rem;
 }
 .Rightside__Notes__BoxandButton {
   width: auto;
